# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A simple Task API"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Users not found"
          }
        }
      }
    },
    "/{email}/todo": {
      "get": {
        "summary": "Get todo tasks by email",
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "description": "User's email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "post": {
        "summary": "Add tasks by email",
        "description": "Add tasks by user's email",
        "tags": ["Task"],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "description": "User's email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Task details to add",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/task"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/{email}/todo/{taskId}": {
      "get": {
        "summary": "Get a specific todo task by task ID",
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "description": "User's email",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "taskId",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "User or task not found"
          }
        }
      },
      "put": {
        "summary": "Update a specific todo task by task ID",
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "description": "User's email",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "taskId",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Updated task details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/task"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid task details"
          },
          "404": {
            "description": "User or task not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific task for a user by taskID",
        "description": "Deletes a todo task for the specified user using their email and the task's ID.",
        "tags": ["Task"],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "description": "User's email",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "taskId",
            "description": "Task's unique ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Task successfully deleted"
          },
          "404": {
            "description": "User or Task not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "task": {
        "type": "object",
        "properties": {
          "task": {
            "type": "string"
          },
          "completed": {
            "type": "boolean"
          }
        }
      }
    }
  }
}